update 7 june

+++ resources/views/showAllProducts.blade.php
@@ -1,6 +1,21 @@
 @extends('layouts.app')
 
 @section('content')
+<script>
+    var deleteLinks = document.querySelectorAll('.delete');
+
+    for (var i = 0; i < deleteLinks.length; i++) {
+        deleteLinks[i].addEventListener('click', function(event) {
+            event.preventDefault();
+
+            var choice = confirm(this.getAttribute('data-confirm'));
+
+            if (choice) {
+                window.location.href = this.getAttribute('href');
+            }
+        });
+    }
+</script>
     <div class="container">
         @if (count($errors) > 0)
             <div class="alert alert-danger">
@@ -19,7 +34,33 @@
             </div>
         @endif
         <div class="row ">
+            <div class="row justify-content-center">
+<div class="col-lg-6">
+                    <div class="card">
+<div class="col-lg-10">
+                        <div class="card-header">Search Product</div>
+                        <form method="GET" action="/home/searchProductName1/{{$site_id}}">
 
+
+                            <input type="text" class="col-6" id="productName"  name="productName" placeholder="enter the product name">
+
+
+                            <button type="submit" class="btn btn-default"  >search</button>
+                        </form>
+</div>
+                        <div class="col-lg-4">
+                            <select class="form-control" onchange="this.options[this.selectedIndex].value && (window.location = this.options[this.selectedIndex].value);">
+                                <option>sort list</option>
+                                <option >A-Z</option>
+                                <option>Z-A</option>
+                                <option>Low-High</option>
+                                <option>High-Low</option>
+                            </select>
+                        </div>
+
+
+                </div>
+</div>
                 <div class="card">
                     <div class="card-header">All Products</div>
                     <table class="table table-hover">
@@ -37,6 +78,23 @@
                         </tr>
                         </thead>
                         <tbody>
+                        @if(isset($_GET['productName']))
+
+                            @foreach($products as $singleProduct)
+                                <tr>
+                                    <td>{{$singleProduct->id}}</td>
+                                    <td>{{$singleProduct->site_id}}</td>
+                                    <td>{{$singleProduct->product_name}}</td>
+                                    <td>{{$singleProduct->product_description}}</td>
+                                    <td>{{$singleProduct->price}}</td>
+                                    <td>{{$singleProduct->barcode}}</td>
+                                    <td>{{$singleProduct->quantity}}</td>
+                                    <td><a class="btn btn-danger "  href="">Delete</a></td>
+                                    <td><a class="btn btn-success" href="/home/updateProductInfo/{{$singleProduct->id}}/{{$site_id}}">Update</a></td>
+                                </tr>
+                            @endforeach
+
+                        @else
                         @foreach($products as $singleProduct)
                             <tr>
                                 <td>{{$singleProduct->id}}</td>
@@ -46,14 +104,19 @@
                                 <td>{{$singleProduct->price}}</td>
                                 <td>{{$singleProduct->barcode}}</td>
                                 <td>{{$singleProduct->quantity}}</td>
-                                <td><a class="btn btn-danger" href="/deleteProduct/{{$singleProduct->id}}">Delete</a></td>
-                                <td><a class="btn btn-success" href="/updateProductInfo/{{$singleProduct->id}}">Update</a></td>
+                                <td><a class="btn btn-danger confirm"   href="/home/deleteProduct/{{$singleProduct->id}}">Delete</a></td>
+                                <td><a class="btn btn-success" href="/home/updateProductInfo/{{$singleProduct->id}}/{{$site_id}}">Update</a></td>
                             </tr>
                         @endforeach
                         </tbody>
+                            @endif
                     </table>
+                    {{ $products->links() }}
                 </div>
 
         </div>
     </div>
+    </div>
+
+
 @endsection
